Leave only the search itself in unit-search
The pool, the costs, the styles and the names of things are what the site has
decided, and they belong with the quantities they describe. What is left in
unit-search is the procedure: cover the dimensions with units from a pool that
is handed to it, count what each covering costs, and return the cheapest.
